+++ ProjectProgPtrns/src/main/java/org/gigi/model/Book.java
@@ -48,6 +48,11 @@ public abstract class Book {
         }
     }
 
+    // todo
+    public boolean isAvailable() {
+        return true;
+    }
+
 
     /**
      * method that checks if the name is not valid(has numbers or special charcters)
