+++ ProjectProgPtrns/src/main/java/org/gigi/model/Student.java
@@ -2,7 +2,6 @@ package org.gigi.model;
 
 
 import lombok.ToString;
-
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -15,96 +14,20 @@ import java.util.List;
 @Getter
 @Setter
 public class Student extends User {
-
-    private static int nextId = 1;
     private static final int MAX_BOOKS_ALLOWED = 3;
 
     public Student(String firstName, String lastName, String email) {
-        super(nextId++, firstName, lastName, email,MAX_BOOKS_ALLOWED);
+        super(firstName, lastName, email, MAX_BOOKS_ALLOWED);
     }
-
-    public void borrowBook(String name) {
-
-    }
-
 
     @Override
     public String getDetails() {
-        return "Student Details:" + "\nId: " + this.userId +
-                "\nFirst Name: " + firstName +
-                "\nLast Name: " + lastName +
-                "\nEmail: " + email +
-                "\nOverdue Books: " + getOverdueRecords() +
-                "\nIssued Books: " + issuedBooks.size();
+        return String.format(
+                "Student Details:\nId: %d\nName: %s\nEmail: %s\nOverdue Books: %d\nIssued Books: %d",
+                userId, getFullName(), email, getOverdueRecords().size(), issuedBooks.size()
+        );
     }
 
 
-    @Override
-    public List<BorrowedBookRecord> getOverdueRecords() {
-        List<BorrowedBookRecord> overdueRecords = new ArrayList<>();
-        for (BorrowedBookRecord record : issuedBooks) {
-            if (record.isOverDue()) {
-                overdueRecords.add(record);
-            }
-        }
-        return overdueRecords;
-    }
-
-    @Override
-    public boolean canBorrowBook(Book book) { //gon look at this more deeply later
-        if (issuedBooks.size() <= MAX_BOOKS_ALLOWED) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    @Override
-    public boolean returnBook(Book book) {
-        List<BorrowedBookRecord> borrowedBookRecords = LibrarySystem.getInstance().getBorrowedBookRecords();
-
-        for (BorrowedBookRecord record : issuedBooks) {
-            if (record.getBook().equals(book)) {
-                record.getBook().incrementAvailableCopies();
-                issuedBooks.remove(record);
-                borrowedBookRecords.remove(record);
-
-                System.out.println("Book returned successfully: " + record.getBook().getTitle());
-                return true;
-            }
-        }
-
-        System.out.println("No record found for book with ISBN " + book.getTitle());
-        return false;
-    }
-
-    public boolean returnBook(String bookISBN) {
-        List<BorrowedBookRecord> borrowedBookRecords = LibrarySystem.getInstance().getBorrowedBookRecords();
-        for (BorrowedBookRecord record : issuedBooks) {
-            if (record.getBook().getIsbn().equals(bookISBN)) {
-                record.getBook().incrementAvailableCopies();
-                borrowedBookRecords.remove(record);
-                System.out.println("Book returned successfully: " + record.getBook().getTitle());
-                return true;
-            }
-        }
-        System.out.println("No record found for book with ISBN " + bookISBN);
-        return false;
-    }
-
-    @Override
-    public List<Book> searchBook(String keyword) {
-        List<Book> searchResults = new ArrayList<>();
-        List<Book> books = LibrarySystem.getInstance().getBooks();
-
-        for (Book book : books) {
-            if (book.getTitle().toLowerCase().contains(keyword.toLowerCase()) ||
-                    book.getAuthorFName().toLowerCase().contains(keyword.toLowerCase()) ||
-                    book.getAuthorLName().toLowerCase().contains(keyword.toLowerCase())) {
-                searchResults.add(book);
-            }
-        }
-        return searchResults;
-    }
 }
